fix: run on web or app simulataneouly
