--- KakaoLoginExample/App.tsx
@@ -1,27 +1,0 @@
-/**
- * Sample React Native App
- * https://github.com/facebook/react-native
- *
- * Generated with the TypeScript template
- * https://github.com/react-native-community/react-native-template-typescript
- *
- * @format
- */
-
-import React from 'react';
-import {
-  SafeAreaView,
-} from 'react-native';
-import Intro from './src/native/Intro';
-
-
-const App = () => {
-
-  return (
-    <SafeAreaView >
-      <Intro />
-    </SafeAreaView>
-  );
-};
-
-export default App;
