+++ KakaoLoginExample/__tests__/App-test.tsx
@@ -3,9 +3,9 @@
  */
 
 import 'react-native';
-import React from 'react';
-import App from '../App';
 
+import App from '../src/App';
+import React from 'react';
 // Note: test renderer must be required after react-native.
 import renderer from 'react-test-renderer';
