+++ KakaoLoginExample/src/App.tsx
@@ -1,15 +1,26 @@
-import React from 'react';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
-import Login from './pages/IntroWeb';
+/**
+ * Sample React Native App
+ * https://github.com/facebook/react-native
+ *
+ * Generated with the TypeScript template
+ * https://github.com/react-native-community/react-native-template-typescript
+ *
+ * @format
+ */
 
-function App() {
+import Intro from './pages/Intro';
+import React from 'react';
+import {
+  SafeAreaView,
+} from 'react-native';
+
+const App = () => {
+
   return (
-    <BrowserRouter>
-      <Routes>
-        <Route path="/" element={<Login />} />
-      </Routes>
-    </BrowserRouter>
+    <SafeAreaView >
+      <Intro />
+    </SafeAreaView>
   );
-}
+};
 
 export default App;
