+++ KakaoLoginExample/src/index.js
@@ -1,5 +1,5 @@
-import { AppRegistry } from "react-native";
 import App from "./App";
+import { AppRegistry } from "react-native";
 
 const appName = "KakaoLoginWeb";
